Release of version 0.6.0 featuring a rebuild uTP implementation

+++ bittorrent/src/main/java/org/johnnei/javatorrent/utils/MathUtils.java
@@ -10,7 +10,7 @@ public class MathUtils {
 		return (value + (divisor - 1)) / divisor;
 	}
 
-	public static long ceilDivision(long value, int divisor) {
+	public static long ceilDivision(long value, long divisor) {
 		return (value + (divisor - 1)) / divisor;
 	}
